requirements.txt
functions-framework==3.*
google-cloud-firestore==2.5.0
flask_cors==3.0.10